movendo componentes do billing
